Rewrite plugin architecture
The plugin architecture has been rewritten so that there
is now one thread for each plugin and not for each object
any more. The objects are now processed in a queue and
the results are stored in a write cache module that writes
data to the RRD databases regularly or when needed.

This enables us to achieve our goal for a zeroconf approach
much better because every time data is collected it is scanned
for the available disks or interfaces or what ever is monitored
again.