latency: Rewrite latency module
This patch replaces the builtin python implementation
that pinged hosts by a Python C module that uses liboping.

liboping is able to ping IPv6 hosts as well and should
implement the ICMP protocol more appropriately.

The graph has been extended so that hosts will have a
line for latency over IPv6 and IPv4 if available and
the packet loss is merged from both protocols, too.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>