Add graph info functionality
The GraphInfo interface will return some basic information about
a graph (like title, etc.) and can be extended in the future.

This patch also changed the return value of the GenerateGraph
interface which is now a dictionary which is more extensible
and comes with some metrics about the graph now.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>