Drop entropy plugin
This is no longer useful.

The Linux kernel has changed to only use the its internal PRNG
for /dev/random and /dev/urandom.

Once it is seeded, it will spit out as many random data as you
would like, but would not require an entropy pool being filled
up as much.

Signed-off-by: Michael Tremer <michael.tremer@ipfire.org>